refactor(parser): split stmt/assign.rs into cohesive submodules (§7-8)
Split the 2573-line parser/stmt/assign.rs (over the 500-line limit) into a
facade plus nine themed submodules under src/parser/stmt/assign/. All but
one are under 500 lines; assign_stmt.rs (520) holds the single indivisible
assign_stmt function. assign.rs is a hybrid of the CompoundAssignOp enum +
impl and free parser functions addressed as assign::<name> (59 call sites);
the facade keeps the enum + TMP_INDEX_COUNTER and re-exports each submodule
preserving its original visibility, so no external caller changes.
stmt/mod.rs is untouched.

- assign.rs (96)         - facade: enum CompoundAssignOp, counter, re-exports
- assign_stmt.rs (520)   - assign_stmt (single indivisible fn)
- try_assign.rs (462)    - try_parse_assign_expr + colon method arg
- paren.rs (398)         - parenthesized assignment parsing
- op.rs (396)            - CompoundAssignOp impl + compound-op name parsing
- compound_expr.rs (326) - build_compound/custom/meta assign expr
- lvalue.rs (147)        - method/named-sub/callable/subscript/list lvalue builders
- comma.rs (102)         - parse_comma_or_expr + lift_meta_ops_in_list
- bracket.rs (99)        - parse_bracket_meta_assign_op
- sink.rs (79)           - parse_assign_expr_or_comma + sink rewrites

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; cross-submodule private helpers bumped to
pub(crate); two never-external items reduced to module-private. One path
qualified to absolute. The fn multiset is byte-identical. No behavior change.
